+++ src/main/java/com/elina/SchoolSpringJavaNoSD/service/CourseService.java
@@ -8,7 +8,7 @@ public interface CourseService {
 
     public void saveCourse(Course theCourse);
 
-    public Course getCourse(int theId);
+    public Course getCourse(Long theId);
 
-    public void deleteCourse(int theId);
+    public void deleteCourse(Long theId);
 }
